An AP value was being recast to INT8 which caused some issues with more then 127APs.
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@2441 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Tactical/Points.cpp
@@ -616,7 +616,7 @@ void DeductPoints( SOLDIERTYPE *pSoldier, INT16 sAPCost, INT32 iBPCost,BOOLEAN f
 		sNewAP = 0;
 	}
 
-	pSoldier->bActionPoints = (INT8)sNewAP;
+	pSoldier->bActionPoints = sNewAP;
 
 	DebugMsg( TOPIC_JA2, DBG_LEVEL_3, String("Deduct Points (%d at %d) %d %d", pSoldier->ubID, pSoldier->sGridNo, sAPCost, iBPCost	) );
